# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CareerPilot \u2014 Multi-agent AI job discovery pipeline",
    "description": "Three agents (Scout finds company job boards, Analyst scores postings against my resume, Coach identifies recurring skill gaps), LangGraph orchestration, RAG over the job archive, and a hand-labeled evaluation set.\n\nThis is the HTTP layer over that pipeline. It wraps adapters/, filters.py, agents/, orchestrator.py, db.py, extraction.py and rag.py - it does not reimplement any of them.\n\n**The system finds and ranks; the human applies.** Nothing here submits an application to an employer, and no endpoint exists that could - that is a design decision, not a missing feature.\n\nRanking quality is measured against the hand-labeled set rather than assumed. The current result is marginal - stage-1 MRR sits only slightly above the random baseline on a small overlap. See GET /api/meta/evaluation for the numbers and their caveats.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Jobs",
        "description": "Ranked survivors, best fit first. Unscored jobs (the Analyst found\nno concrete requirements to compare against) are appended after the\nscored ones rather than sorted among them - their fit_score is not a\nreal comparison and must not compete for a ranking position.",
        "operationId": "list_jobs_api_jobs_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "filter by application_status; repeatable",
              "title": "Status"
            },
            "description": "filter by application_status; repeatable"
          },
          {
            "name": "include_unscored",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Include Unscored"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobSummary"
                  },
                  "title": "Response List Jobs Api Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/rejected": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Rejected",
        "operationId": "list_rejected_api_jobs_rejected_get",
        "parameters": [
          {
            "name": "rule",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rule"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/jobs/{content_hash}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job",
        "operationId": "get_job_api_jobs__content_hash__get",
        "parameters": [
          {
            "name": "content_hash",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content Hash"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{content_hash}/status": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Update Status",
        "description": "Sets application_status, and applied_at the first time a job is\nmarked \"applied\". This records that the human applied; it never applies\non their behalf - there is no code path from here to an employer.",
        "operationId": "update_status_api_jobs__content_hash__status_post",
        "parameters": [
          {
            "name": "content_hash",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content Hash"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusUpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/preferences": {
      "get": {
        "tags": [
          "preferences"
        ],
        "summary": "Get Preferences",
        "operationId": "get_preferences_api_preferences_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "preferences"
        ],
        "summary": "Update Preferences",
        "operationId": "update_preferences_api_preferences_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferenceLists"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesUpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/preferences/preview": {
      "post": {
        "tags": [
          "preferences"
        ],
        "summary": "Preview Preferences",
        "description": "Impact without saving. Runs the same filter pass the pipeline would,\nagainst the candidate lists, leaving config.* untouched.",
        "operationId": "preview_preferences_api_preferences_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferenceLists"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesImpact"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobDetail": {
        "properties": {
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "company": {
            "type": "string",
            "title": "Company"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "remote_type": {
            "type": "string",
            "title": "Remote Type"
          },
          "fit_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fit Score"
          },
          "verdict": {
            "type": "string",
            "title": "Verdict"
          },
          "is_unscored": {
            "type": "boolean",
            "title": "Is Unscored"
          },
          "matched_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Matched Skills"
          },
          "missing_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Skills"
          },
          "reasoning": {
            "type": "string",
            "title": "Reasoning"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "years_required": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Years Required"
          },
          "resume_meets_experience": {
            "type": "boolean",
            "title": "Resume Meets Experience"
          },
          "application_status": {
            "type": "string",
            "title": "Application Status"
          },
          "applied_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied At"
          },
          "first_seen": {
            "type": "string",
            "format": "date-time",
            "title": "First Seen"
          },
          "is_new": {
            "type": "boolean",
            "title": "Is New"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "content_hash",
          "company",
          "title",
          "url",
          "remote_type",
          "verdict",
          "is_unscored",
          "reasoning",
          "model",
          "resume_meets_experience",
          "application_status",
          "first_seen",
          "is_new",
          "description"
        ],
        "title": "JobDetail"
      },
      "JobSummary": {
        "properties": {
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "company": {
            "type": "string",
            "title": "Company"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "remote_type": {
            "type": "string",
            "title": "Remote Type"
          },
          "fit_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fit Score"
          },
          "verdict": {
            "type": "string",
            "title": "Verdict"
          },
          "is_unscored": {
            "type": "boolean",
            "title": "Is Unscored"
          },
          "matched_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Matched Skills"
          },
          "missing_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Skills"
          },
          "reasoning": {
            "type": "string",
            "title": "Reasoning"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "years_required": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Years Required"
          },
          "resume_meets_experience": {
            "type": "boolean",
            "title": "Resume Meets Experience"
          },
          "application_status": {
            "type": "string",
            "title": "Application Status"
          },
          "applied_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied At"
          },
          "first_seen": {
            "type": "string",
            "format": "date-time",
            "title": "First Seen"
          },
          "is_new": {
            "type": "boolean",
            "title": "Is New"
          }
        },
        "type": "object",
        "required": [
          "content_hash",
          "company",
          "title",
          "url",
          "remote_type",
          "verdict",
          "is_unscored",
          "reasoning",
          "model",
          "resume_meets_experience",
          "application_status",
          "first_seen",
          "is_new"
        ],
        "title": "JobSummary",
        "description": "One ranked survivor. No description - see JobDetail. fit_score and\nverdict are Optional because an \"unscored\" job (no concrete technical\nrequirements in the posting, see agents/analyst.py's is_unscored) has\nno real score, and serving a 0 or a 50 there would be inventing a\njudgment the Analyst explicitly declined to make. is_unscored says\nwhich case a null is."
      },
...
      "PreferencesImpact": {
        "properties": {
          "current_survivors": {
            "type": "integer",
            "title": "Current Survivors"
          },
          "new_survivors": {
            "type": "integer",
            "title": "New Survivors"
          },
          "delta": {
            "type": "integer",
            "title": "Delta"
          },
          "unanalyzed_after": {
            "type": "integer",
            "title": "Unanalyzed After"
          }
        },
        "type": "object",
        "required": [
          "current_survivors",
          "new_survivors",
          "delta",
          "unanalyzed_after"
        ],
        "title": "PreferencesImpact",
        "description": "Survivor counts before and after a candidate edit. `unanalyzed_after`\nis how many of the new survivors have no Analyst result yet - i.e. how\nmuch a run would actually have to do."
      },
      "PreferencesResponse": {
        "properties": {
          "title_allowlist": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Title Allowlist"
          },
          "seniority_keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Seniority Keywords"
          },
          "non_engineering_keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Non Engineering Keywords"
          },
          "india_location_keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "India Location Keywords"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "title_allowlist",
          "seniority_keywords",
          "non_engineering_keywords",
          "india_location_keywords"
        ],
        "title": "PreferencesResponse",
        "description": "`warnings` carries config.load_preferences' fallback notices - an\nemptied or unparseable list means the filters are running on built-in\ndefaults instead of what's on disk, and that has to be visible rather\nthan silently corrected. An empty title_allowlist in particular would\nlet every job through and burn a day's LLM quota on unfiltered noise."
      },
...
      "RejectedPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RejectedJob"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "rules": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rules"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "rules"
        ],
        "title": "RejectedPage"
      },
...
      "StatusUpdateRequest": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "StatusUpdateRequest"
      },
      "StatusUpdateResponse": {
        "properties": {
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "application_status": {
            "type": "string",
            "title": "Application Status"
          },
          "applied_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied At"
          }
        },
        "type": "object",
        "required": [
          "content_hash",
          "application_status"
        ],
        "title": "StatusUpdateResponse"
      },
...
    }
  }
}